+++ Sources/StreamChat/WebSocketClient/Engine/URLSessionWebSocketEngine.swift
@@ -5,12 +5,18 @@
 import Foundation
 
 @available(iOS 13, *)
-class URLSessionWebSocketEngine: NSObject, WebSocketEngine, URLSessionDataDelegate, URLSessionWebSocketDelegate {
-    private var task: URLSessionWebSocketTask?
+class URLSessionWebSocketEngine: NSObject, WebSocketEngine {
+    private weak var task: URLSessionWebSocketTask? {
+        didSet {
+            oldValue?.cancel()
+        }
+    }
     
     let request: URLRequest
-    private var session: URLSession!
+    private var session: URLSession?
     let delegateOperationQueue: OperationQueue
+    let sessionConfiguration: URLSessionConfiguration
+    var urlSessionDelegateHandler: URLSessionDelegateHandler?
     
     var callbackQueue: DispatchQueue { delegateOperationQueue.underlyingQueue! }
     
@@ -18,27 +24,35 @@ class URLSessionWebSocketEngine: NSObject, WebSocketEngine, URLSessionDataDelega
     
     required init(request: URLRequest, sessionConfiguration: URLSessionConfiguration, callbackQueue: DispatchQueue) {
         self.request = request
+        self.sessionConfiguration = sessionConfiguration
         
         delegateOperationQueue = OperationQueue()
         delegateOperationQueue.underlyingQueue = callbackQueue
-        
+
         super.init()
-        
-        session = URLSession(
-            configuration: sessionConfiguration,
-            delegate: self,
-            delegateQueue: delegateOperationQueue
-        )
     }
     
     func connect() {
-        task = session.webSocketTask(with: request)
+        urlSessionDelegateHandler = makeURLSessionDelegateHandler()
+
+        session = URLSession(
+            configuration: sessionConfiguration,
+            delegate: urlSessionDelegateHandler,
+            delegateQueue: delegateOperationQueue
+        )
+
+        task = session?.webSocketTask(with: request)
         doRead()
         task?.resume()
     }
     
     func disconnect() {
         task?.cancel(with: .normalClosure, reason: nil)
+        session?.invalidateAndCancel()
+
+        session = nil
+        task = nil
+        urlSessionDelegateHandler = nil
     }
     
     func sendPing() {
@@ -65,46 +79,75 @@ class URLSessionWebSocketEngine: NSObject, WebSocketEngine, URLSessionDataDelega
             }
         }
     }
+
+    private func makeURLSessionDelegateHandler() -> URLSessionDelegateHandler {
+        let urlSessionDelegateHandler = URLSessionDelegateHandler()
+        urlSessionDelegateHandler.onOpen = { [weak self] _ in
+            self?.callbackQueue.async {
+                self?.delegate?.webSocketDidConnect()
+            }
+        }
+
+        urlSessionDelegateHandler.onClose = { [weak self] closeCode, reason in
+            var error: WebSocketEngineError?
+
+            if let reasonData = reason, let reasonString = String(data: reasonData, encoding: .utf8) {
+                error = WebSocketEngineError(
+                    reason: reasonString,
+                    code: closeCode.rawValue,
+                    engineError: nil
+                )
+            }
+
+            self?.callbackQueue.async { [weak self] in
+                self?.delegate?.webSocketDidDisconnect(error: error)
+            }
+        }
+
+        urlSessionDelegateHandler.onCompletion = { [weak self] error in
+            // If we received this callback because we closed the WS connection
+            // intentionally, `error` param will be `nil`.
+            // Delegate is already informed with `didCloseWith` callback,
+            // so we don't need to call delegate again.
+            guard let error = error else { return }
+
+            self?.callbackQueue.async { [weak self] in
+                self?.delegate?.webSocketDidDisconnect(error: WebSocketEngineError(error: error))
+            }
+        }
+
+        return urlSessionDelegateHandler
+    }
+
+    deinit {
+        disconnect()
+    }
+}
     
+@available(iOS 13, *)
+class URLSessionDelegateHandler: NSObject, URLSessionDataDelegate, URLSessionWebSocketDelegate {
+    var onOpen: ((_ protocol: String?) -> Void)?
+    var onClose: ((_ code: URLSessionWebSocketTask.CloseCode, _ reason: Data?) -> Void)?
+    var onCompletion: ((Error?) -> Void)?
+
     public func urlSession(
         _ session: URLSession,
         webSocketTask: URLSessionWebSocketTask,
         didOpenWithProtocol protocol: String?
     ) {
-        callbackQueue.async { [weak self] in
-            self?.delegate?.webSocketDidConnect()
-        }
+        onOpen?(`protocol`)
     }
-    
+
     func urlSession(
         _ session: URLSession,
         webSocketTask: URLSessionWebSocketTask,
         didCloseWith closeCode: URLSessionWebSocketTask.CloseCode,
         reason: Data?
     ) {
-        var error: WebSocketEngineError?
-        
-        if let reasonData = reason, let reasonString = String(data: reasonData, encoding: .utf8) {
-            error = WebSocketEngineError(
-                reason: reasonString,
-                code: closeCode.rawValue,
-                engineError: nil
-            )
-        }
-        
-        callbackQueue.async { [weak self] in
-            self?.delegate?.webSocketDidDisconnect(error: error)
-        }
+        onClose?(closeCode, reason)
     }
-    
+
     func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
-        // If we received this callback because we closed the WS connection
-        // intentionally, `error` param will be `nil`.
-        // Delegate is already informed with `didCloseWith` callback,
-        // so we don't need to call delegate again.
-        guard let error = error else { return }
-        callbackQueue.async { [weak self] in
-            self?.delegate?.webSocketDidDisconnect(error: WebSocketEngineError(error: error))
-        }
+        onCompletion?(error)
     }
 }