+++ Sources/StreamChat/Workers/UserListUpdater.swift
@@ -23,10 +23,10 @@ class UserListUpdater: Worker {
     ///
     func update(userListQuery: UserListQuery, policy: UpdatePolicy = .merge, completion: ((Error?) -> Void)? = nil) {
         apiClient
-            .request(endpoint: .users(query: userListQuery)) { (result: Result<UserListPayload, Error>) in
+            .request(endpoint: .users(query: userListQuery)) { [weak self] (result: Result<UserListPayload, Error>) in
                 switch result {
                 case let .success(userListPayload):
-                    self.database.write { session in
+                    self?.database.write { session in
                         if case .replace = policy {
                             let dto = try session.saveQuery(query: userListQuery)
                             dto?.users.removeAll()